add : ILPを行うServiceクラスと、結果保持専用のクラスを作成

+++ backend/src/main/java/com/gwork/demo/testGenetic.java
@@ -23,8 +23,8 @@ public class testGenetic {
   static double[][] stapleAndProtein = nutrientService.getStapleAndProtein(); //主食・肉の栄養テーブル
   static double[][] vegetable = nutrientService.getVegetable(); //野菜類の栄養テーブル
   static double[] prices = setPrices(nutrientService.getPriceUnit(), jsonProcesserService.getIngAndPri()); //100gあたりの野菜類の価格情報
-  static int[] staVolOfVeg = nutrientService.getStaVolOfVeg(); //野菜類の1食分の最低量
-  static int[] staVolOfsAndP = nutrientService.getStaVolOfsAndP(); //主食・肉類の1食分の目安量
+  static double[] staVolOfVeg = nutrientService.getStaVolOfVeg(); //野菜類の1食分の最低量
+  static double[] staVolOfsAndP = nutrientService.getStaVolOfsAndP(); //主食・肉類の1食分の目安量
   static double[] targets = nutrientService.getTargets(); //目標値のテーブル
   
   // --- メインループ ---
